# api/generated/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Send and receive Messages",
    "version" : "1.0.0",
    "title" : "messagehub"
  },
  "host" : "localhost:8080",
  "basePath" : "/api",
  "schemes" : [ "http", "https" ],
  "paths" : {
    "/" : {
      "post" : {
        "summary" : "Retrieves all comments of a message by messageId",
        "description" : "",
        "operationId" : "addComment",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "messageId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/messages" : {
      "get" : {
        "summary" : "Retrieves all messages, (?start=x ?size=x&year=xxxx or paginated only ?start=x&size=x)",
        "description" : "",
        "operationId" : "getMessages",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "year",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "start",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "Ok"
          },
          "204" : {
            "description" : "No Messages available yet"
          }
        }
      }
    },
    "/messages/{messageId}" : {
      "get" : {
        "summary" : "Retrieves a single message by it's id",
        "description" : "",
        "operationId" : "getMessage",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "messageId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Ok"
          },
          "404" : {
            "description" : "No Message with given id was found"
          }
        }
      },
      "put" : {
        "summary" : "Updates a single message by it's id",
        "description" : "",
        "operationId" : "updateMessage",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "messageId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Ok, message updated with new text"
          },
          "404" : {
            "description" : "No message with given id was found"
          }
        }
      },
      "delete" : {
        "summary" : "Deletes a single message",
        "description" : "",
        "operationId" : "deleteMessage",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "messageId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "204" : {
            "description" : "Delete successful, no content in return"
          },
          "404" : {
            "description" : "Nothing found to delete"
          }
        }
      }
    },
    "/messages/{profileId}" : {
      "post" : {
        "summary" : "Excepts a new message",
        "description" : "",
        "operationId" : "addMessage",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "profileId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/profiles" : {
      "get" : {
        "summary" : "Retrieves all profiles",
        "description" : "",
        "operationId" : "getProfiles",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "Ok"
          },
          "204" : {
            "description" : "No profiles found yet"
          }
        }
      },
      "post" : {
        "summary" : "Excepts a new profile",
        "description" : "",
        "operationId" : "addProfile",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/profiles/{profileId}" : {
      "get" : {
        "summary" : "Retrieves a profile by profileId",
        "description" : "",
        "operationId" : "getProfile",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "profileId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Ok"
          },
          "404" : {
            "description" : "No profile with given id was found"
          }
        }
      },
      "put" : {
        "summary" : "Updates a profile by it's given id",
        "description" : "",
        "operationId" : "updateProfile",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "profileId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Ok, profile updated with new text"
          },
          "404" : {
            "description" : "No profile with given id was found"
          }
        }
      },
      "delete" : {
        "summary" : "Deletes a profile by profileId",
        "description" : "",
        "operationId" : "deleteProfile",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "profileId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "204" : {
            "description" : "Delete successful, no content in return"
          },
          "404" : {
            "description" : "Nothing found to delete"
          }
        }
      }
    }
  }
}